Module Restructure and TODOs (#7)
