+++ Makefile
@@ -7,11 +7,44 @@ else
 mm = 
 endif 
 
+C_SCRIPT_SRC = plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks.c
+C_SCRIPT_AARCH64 = plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks_linux_aarch64
+C_SCRIPT_X86_64 = plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks_linux_x86_64
+C_SCRIPT_ALL_TARGETS = $(C_SCRIPT_X86_64) $(C_SCRIPT_AARCH64)
+
+arch:=$(shell uname -m)
+
+ifneq ($(arch),x86_64)
+$(info ====================================================================================)
+$(info This Makefile assumes to be run on an x86_64 build system. Found: $(arch))
+$(info As an alternative you can adjust the Makefile or build $(C_SCRIPT_SRC) yourself for the following targets:)
+$(info - target aarch64-linux-gnu: saved to $(C_SCRIPT_AARCH64))
+$(info - target x86_64-linux-gnu: saved to $(C_SCRIPT_X86_64))
+$(error Aborting)
+endif
+
+$(C_SCRIPT_AARCH64): $(C_SRIPT_SRC)
+	clang -static \
+		-target aarch64-linux-gnu \
+		--sysroot=/usr/aarch64-linux-gnu \
+		plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks.c \
+		-fuse-ld=lld \
+		-o $@
+
+
+$(C_SCRIPT_X86_64): $(C_SCRIPT_SRC)
+	clang -static \
+		-target x86_64-linux-gnu \
+		plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks.c \
+		-fuse-ld=lld \
+		-o $@
+
 clean:
 	rm -rf .nextflow*
 	rm -rf work
 	rm -rf build
 	rm -rf plugins/*/build
+	rm $(C_SCRIPT_ALL_TARGETS)
 	./gradlew clean
 
 compile:
@@ -55,10 +88,7 @@ assemble:
 # generate build zips under build/plugins
 # you can install the plugin copying manually these files to $HOME/.nextflow/plugins
 #
-buildPlugins:
-	# TODO: add targets
-	# clang -static -target arm-linux-gnu plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks.c -o plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks_linux_aarch86
-	clang -static -target x86_64-linux-gnu plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks.c -o plugins/nf-cws/src/resources/nf-cws/getStatsAndResolveSymlinks_linux_x86
+buildPlugins: $(C_SCRIPT_ALL_TARGETS)
 	./gradlew copyPluginZip
 
 #
