+++ plugins/nf-cws/src/main/nextflow/cws/CWSPlugin.groovy
@@ -1,9 +1,11 @@
 package nextflow.cws
 
 import groovy.transform.CompileStatic
+import nextflow.cws.wow.file.LocalPath
 import nextflow.cws.wow.file.OfflineLocalPath
-import nextflow.cws.wow.file.WOWFileSystemProvider
-import nextflow.cws.wow.file.WorkdirHelper
+import nextflow.cws.wow.file.WorkdirPath
+import nextflow.cws.wow.filesystem.WOWFileSystemProvider
+import nextflow.cws.wow.serializer.LocalPathSerializer
 import nextflow.file.FileHelper
 import nextflow.plugin.BasePlugin
 import nextflow.trace.TraceRecord
@@ -62,8 +64,9 @@ class CWSPlugin extends BasePlugin {
     void start() {
         super.start()
         registerTraceFields()
-        KryoHelper.register( OfflineLocalPath )
-        KryoHelper.register( WorkdirHelper )
+        KryoHelper.register( LocalPath, LocalPathSerializer )
+        KryoHelper.register( OfflineLocalPath, LocalPathSerializer )
+        KryoHelper.register( WorkdirPath, LocalPathSerializer )
         FileHelper.getOrInstallProvider(WOWFileSystemProvider)
     }
