+++ plugins/nf-cws/src/main/nextflow/cws/k8s/K8sSchedulerClient.groovy
@@ -4,7 +4,6 @@ import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import nextflow.cws.CWSConfig
 import nextflow.cws.SchedulerClient
-import nextflow.cws.wow.file.LocalPath
 import nextflow.exception.NodeTerminationException
 import nextflow.k8s.K8sConfig
 import nextflow.k8s.client.K8sResponseException
@@ -20,11 +19,17 @@ import java.nio.file.Paths
 class K8sSchedulerClient extends SchedulerClient {
 
     private final CWSK8sConfig.K8sScheduler schedulerConfig
+
     private final CWSK8sClient k8sClient
+
     private final K8sConfig k8sConfig
+
     private final String namespace
+
     private final Collection<PodHostMount> hostMounts
+
     private final Collection<PodVolumeClaim> volumeClaims
+
     private String ip
 
     K8sSchedulerClient(
@@ -44,7 +49,6 @@ class K8sSchedulerClient extends SchedulerClient {
         this.k8sConfig = k8sConfig
         this.schedulerConfig = schedulerConfig
         this.namespace = namespace ?: 'default'
-        LocalPath.setClient(this)
     }
 
     protected String getDNS(){
