+++ plugins/nf-cws/src/main/nextflow/cws/wow/file/LocalFile.groovy
@@ -18,4 +18,5 @@ class LocalFile extends File {
     Path toPath() {
         return localPath
     }
+
 }
