--- plugins/nf-cws/src/main/nextflow/cws/wow/file/WOWFileSystem.groovy
@@ -1,80 +1,0 @@
-package nextflow.cws.wow.file
-
-import groovy.transform.CompileDynamic
-import groovy.transform.CompileStatic
-import java.nio.file.*
-import java.nio.file.attribute.UserPrincipalLookupService
-import java.nio.file.spi.FileSystemProvider
-
-@CompileStatic
-class WOWFileSystem extends FileSystem {
-
-    static final WOWFileSystem INSTANCE = new WOWFileSystem()
-
-    @Override
-    FileSystemProvider provider() {
-        return WOWFileSystemProvider.INSTANCE
-    }
-
-    @Override
-    void close() throws IOException {
-    }
-
-    @Override
-    boolean isOpen() {
-        return true
-    }
-
-    @Override
-    boolean isReadOnly() {
-        return false
-    }
-
-    @Override
-    String getSeparator() {
-        return "/"
-    }
-
-    @Override
-    Iterable<Path> getRootDirectories() {
-        throw new UnsupportedOperationException("Root directories not supported by ${provider().getScheme().toUpperCase()} file system")
-    }
-
-    @Override
-    Iterable<FileStore> getFileStores() {
-        throw new UnsupportedOperationException("File stores not supported by ${provider().getScheme().toUpperCase()} file system")
-    }
-
-    @Override
-    Set<String> supportedFileAttributeViews() {
-        return new HashSet<>()
-    }
-
-    @Override
-    Path getPath(String s, String... strings) {
-        throw new UnsupportedOperationException("Path get not supported by ${provider().getScheme().toUpperCase()} file system")
-    }
-
-    @Override
-    @CompileDynamic
-    PathMatcher getPathMatcher(String s) {
-        return new PathMatcher() {
-            private final def matcher = FileSystems.getDefault().getPathMatcher( s )
-            @Override
-            boolean matches(Path path) {
-                // Make this a Unix Path and use the default matcher
-                return matcher.matches( Path.of(path.toString()) )
-            }
-        }
-    }
-
-    @Override
-    UserPrincipalLookupService getUserPrincipalLookupService() {
-        throw new UnsupportedOperationException("User principal lookup service not supported by ${provider().getScheme().toUpperCase()} file system")
-    }
-
-    @Override
-    WatchService newWatchService() throws IOException {
-        throw new UnsupportedOperationException("Watch service not supported by ${provider().getScheme().toUpperCase()} file system")
-    }
-}
