+++ plugins/nf-cws/src/main/nextflow/cws/wow/filesystem/WOWOutputStream.groovy
@@ -1,13 +1,16 @@
-package nextflow.cws.wow.file
+package nextflow.cws.wow.filesystem
 
 import groovy.transform.CompileStatic
 import nextflow.cws.SchedulerClient
+import nextflow.cws.wow.file.LocalPath
 
 @CompileStatic
 class WOWOutputStream extends OutputStream {
 
     private OutputStream inner
+
     private SchedulerClient client
+
     private LocalPath path
 
     WOWOutputStream(OutputStream inner, SchedulerClient client, LocalPath path) {
@@ -25,8 +28,9 @@ class WOWOutputStream extends OutputStream {
     @Override
     void close() throws IOException {
         inner.close()
-        Map location = path.getLocation()
+        Map location = WOWFileSystemProvider.INSTANCE.getLocation( path )
         File file = path.getInner().toFile()
         client.addFileLocation(path.toString(), file.size(), file.lastModified(), location.locationWrapperID as long, true)
     }
+
 }
