+++ plugins/nf-cws/src/resources/META-INF/extensions.idx
@@ -1,3 +1,3 @@
 nextflow.cws.CWSFactory
 nextflow.cws.k8s.CWSK8sExecutor
-nextflow.cws.wow.file.WOWFileSystemPathFactory
+nextflow.cws.wow.filesystem.WOWFileSystemPathFactory
